rename `QuizAttemptComponent`
it's now also used for `QuizView`
