+++ src/app/components/elements/quiz/QuizContentsComponent.tsx
@@ -52,7 +52,7 @@ export interface QuizAttemptProps extends QuizProps {
     questions: QuestionDTO[];
     sections: { [id: string]: IsaacQuizSectionDTO };
 }
-export interface QuizViewProps extends QuizProps {
+interface QuizViewProps extends QuizProps {
     attempt?: undefined;
     view: QuizView;
     preview?: undefined;
@@ -66,8 +66,8 @@ function inSection(section: IsaacQuizSectionDTO, questions: QuestionDTO[]) {
     return questions.filter(q => q.id?.startsWith(section.id as string + "|"));
 }
 
-function QuizContents({attempt, sections, questions, pageLink}: QuizAttemptProps) {
-    if (isDefined(attempt?.completedDate)) {
+function QuizDetails({attempt, sections, questions, pageLink}: QuizAttemptProps) {
+    if (isDefined(attempt.completedDate)) {
         return attempt.feedbackMode === "NONE" ?
             <h4>No feedback available</h4>
             : attempt.feedbackMode === "OVERALL_MARK" ?
@@ -125,7 +125,7 @@ function QuizHeader({attempt, preview, view, user}: QuizAttemptProps | QuizViewP
     if (preview || view) {
         const quiz = view ? view.quiz : attempt.quiz;
         return <>
-            {preview && <EditContentButton doc={attempt?.quiz} />}
+            {preview && <EditContentButton doc={attempt.quiz} />}
             <div data-testid="quiz-action" className="d-flex">
                 <p>{ preview ? "You are previewing this test." : "You are viewing the rubric for this test."}</p>
                 <Spacer />
@@ -184,7 +184,7 @@ function QuizSection({attempt, page, studentUser, user, quizAssignmentId}: QuizA
     const openQuestionModal = (attempt: QuizAttemptDTO) => {
         dispatch(openActiveModal({
             closeAction: () => {dispatch(closeActiveModal());}, size: "lg",
-            title: "Test Instructions", body: <QuizRubric attempt={attempt}/>
+            title: "Test Instructions", body: <QuizRubric attempt={attempt} />
         }));
     };
 
@@ -235,7 +235,7 @@ const getCrumbs = (preview: boolean | undefined, view: boolean | undefined, user
 };
 
 const QuizTitle = ({attempt, view, page, pageLink, pageHelp, preview, studentUser, user}: QuizAttemptProps | QuizViewProps) => {
-    const quiz = attempt?.quiz ?? view?.quiz;
+    const quiz = attempt ? attempt.quiz : view.quiz;
     let quizTitle = quiz?.title || quiz?.id || "Test";
     if (isDefined(attempt?.completedDate)) {
         quizTitle += " Feedback";
@@ -290,7 +290,7 @@ function QuizOverview(props: QuizAttemptProps | QuizViewProps) {
             You are viewing this test as <b>{studentUser?.givenName} {studentUser?.familyName}</b>.{quizAssignmentId && <> <Link to={`/test/assignment/${quizAssignmentId}/feedback`}>Click here</Link> to return to the teacher test feedback page.</>}
         </div>}
         <QuizRubric {...props}/>
-        { props.attempt && <QuizContents {...props} /> }
+        { props.attempt && <QuizDetails {...props} /> }
     </div>;
 }
 
@@ -304,7 +304,7 @@ function QuizQuestions(props: Omit<QuizAttemptProps, 'page'> & {page: number}) {
     </QuizAttemptContext.Provider>;
 }
 
-export function QuizAttemptComponent(props: QuizAttemptProps | QuizViewProps) {
+export function QuizContentsComponent(props: QuizAttemptProps | QuizViewProps) {
     return <>
         <QuizTitle {...props} />
         {props.page === null || props.page === undefined ? <QuizOverview {...props}/> : <QuizQuestions {...props} page={props.page} />}
