+++ src/app/components/pages/quizzes/QuizView.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import {useGetQuizRubricQuery} from "../../../state";
 import {Link, useParams} from "react-router-dom";
 import {isTeacherOrAbove, siteSpecific, tags} from "../../../services";
-import {QuizAttemptComponent, rubricCrumbs} from "../../elements/quiz/QuizAttemptComponent";
+import {QuizContentsComponent, rubricCrumbs} from "../../elements/quiz/QuizContentsComponent";
 import {Button, Col, Container, Row} from "reactstrap";
 import {ShowLoadingQuery} from "../../handlers/ShowLoadingQuery";
 import type { RegisteredUserDTO } from "../../../../IsaacApiTypes";
@@ -33,14 +33,14 @@ const QuizFooter = ({quizId, user}: {quizId: string, user: RegisteredUserDTO}) =
 export const QuizView = ({user}: {user: RegisteredUserDTO}) => {
     const {quizId} = useParams<{quizId: string}>();
     const quizRubricQuery = useGetQuizRubricQuery(quizId);
-    const attempt = {
+    const view = {
         quiz: quizRubricQuery.data && tags.augmentDocWithSubject(quizRubricQuery.data),
         quizId: quizRubricQuery.data?.id,
     };
 
-    return <Container className={`mb-5 ${attempt?.quiz?.subjectId}`}>
+    return <Container className={`mb-5 ${view?.quiz?.subjectId}`}>
         <ShowLoadingQuery query={quizRubricQuery} ifError={Error}>
-            <QuizAttemptComponent view={attempt} pageHelp={pageHelp} user={user} />
+            <QuizContentsComponent view={view} pageHelp={pageHelp} user={user} />
             <QuizFooter quizId={quizId} user={user} />
         </ShowLoadingQuery>
     </Container>;
